+++ app/Widgets/Stories/stories.js
@@ -1,0 +1,3 @@
+MovimWebsocket.attach(() => {
+    Stories_ajaxHttpGet();
+});